Rename error file to "errors"
