+++ packages/record/src/errors.ts
