+++ packages/record/src/index.spec.ts
@@ -1,6 +1,6 @@
 import { LookitWindow } from "@lookit/data/dist/types";
 import { initJsPsych } from "jspsych";
-import { ExistingRecordingError, NoSessionRecordingError } from "./error";
+import { ExistingRecordingError, NoSessionRecordingError } from "./errors";
 import Rec from "./index";
 import Recorder from "./recorder";
